refactor: separate and create files routes
Create files index-route and product-route
