+++ bin/server.js
@@ -4,15 +4,10 @@ const app = require('../src/app');
 const http = require('http');
 const debug = require('debug')('nodestr:server');
 
-
-
 const port = normalizePort(process.env.PORT || '3000');
 app.set('port', port);
 
 const server = http.createServer(app);
-
-
-
 
 server.listen(port);
 server.on('error', onError);
