+++ src/app.js
@@ -2,39 +2,14 @@
 const bodyParser = require('body-parser');
 const express = require('express');
 const app = express();
-const router = express.Router();
+
+const indexRoute = require('./routes/index-route');
+const productRoute = require('./routes/product-route');
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false}));
 
-const route = router.get('/', (req, res, next) => {
-    res.status(200).send({
-        title: "Node Store API",
-        version: "0.0.1"
-    });
-});
-const create = router.post('/', (req, res, next) => {
-    res.status(201).send(req.body);
-});
-const put = router.put('/:id', (req, res, next) => {
-    const id = req.params.id;
-    res.status(200).send({
-        id: id,
-        item: req.body
-    });
-});
-const del = router.delete('/', (req, res, next) => {
-    res.status(200).send({
-        id: id,
-        item: req.body
-    });
-});
-
-
-
-app.use('/', route);
-app.use('/products', create);
-app.use('/products', put);
-app.use('/products', del);
+app.use('/', indexRoute);
+app.use('/products', productRoute);
 
 module.exports = app;
